// Frontend/src/model/metrics/quality.ts
export interface IQualityGraphDataItem {
  // complexity: number;
  coverage: number;
  duplications: number;
  // issues: number;
  // severity: number;
  maintainability: number;
  projectName: string;
  // qualityGates: number;
  reliability: number;
  security: number;
  // size: number;
  // tests: number;
  teamId: string;
  timestamp: number;
  url: string;
}

export interface IQualityListDataItem {
  // complexity: number;
  coverage: number;
  duplications: number;
  // issues: number;
  // severity: number;
  maintainability: number;
  projectName: string;
  // qualityGates: number;
  reliability: number;
  security: number;
  // size: number;
  // tests: number;
  teamId: string;
  timestamp: number;
  url: string;
}

export interface IQualityDatabaseDataItem {
  // complexity: number;
  coverage: number;
  duplications: number;
  // issues: number;
  // severity: number;
  maintainability: number;
  projectName: string;
  // qualityGates: number;
  reliability: number;
  security: number;
  // size: number;
  // tests: number;
  teamId: string;
  timestamp: number;
  url: string;
}